LayoutTests/inspector/dom/breakpoint-for-event-listener-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Testing DOMAgent.setBreakpointForEventListener and DOMAgent.removeBreakpointForEventListener.

  

== Running test suite: DOM.BreakpointForEventListener
-- Running test case: DOM.setBreakpointForEventListener
Adding breakpoint for event listener...
Clicking button2...
PASS: Should not pause before button2 event handler is run.
Clicking button1...
PASS: Should pause before button1 event handler is run.
PASS: Pause reason should be Listener.
PASS: Pause data eventName should be "click".

-- Running test case: DOM.removeBreakpointForEventListener
Removing breakpoint for event listener...
Clicking button2...
PASS: Should not pause before button2 event handler is run.
Clicking button1...
PASS: Should not pause before button1 event handler is run.

-- Running test case: DOM.setBreakpointForEventListener.Options.Condition

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOM.setBreakpointForEventListener.Options.IgnoreCount

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOM.setBreakpointForEventListener.Options.Action.Log

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOM.setBreakpointForEventListener.Options.Actions.Evaluate

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOM.setBreakpointForEventListener.Invalid
PASS: Should produce an error.
Error: Missing event listener for given eventListenerId

-- Running test case: DOM.removeBreakpointForEventListener.Invalid
PASS: Should produce an error.
Error: Missing event listener for given eventListenerId
